Improvement: Switch to pytest and pytest-ckan.
To follow along with ckanext-scheming, move tests to use pytest.
Using pytest-ckan, this allows us to write newer pytests and have
work in CKAN 2.8 and 2.9.  Tests are not expected to work for
older versions of CKAN installations. At the time of making this
commit this test setup was run against CKAN 2.8 and CKAN 2.9
both running python 2.7. They have not yet been tested against
python 3.

[x] Update readme with pytest run instructions
[x] Update test.ini with with schemas and extra settings to get the tests running (kept getting errors on missing various sections)
[x] Remove legacy nosetest in test_fluent.py
[x] Add test requirements to install pytest and pytest-ckan